pyproject.toml
[tool.poetry]
name = "theoslearningapp"
version = "0.1.0"
description = ""
authors = ["Théophile Madet <theo.madet@posteo.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
black = "^24.4.2"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.1"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
drf-spectacular = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
